fix: err handling
